Fix all Junit tests
