+++ test/setup/repository/insert-email-repository-test-info.sql
@@ -1,2 +1,2 @@
-INSERT INTO soil_and_pimp_live.email_subscriptions 
-     VALUES ('a@a.com', 'b@b.com', 'c@c.com');
+INSERT INTO soil_and_pimp_live.email_subscriptions (email_address)
+     VALUES ('a@a.com'), ('b@b.com'), ('c@c.com');
