+++ test/setup/repository/insert-event-repository-test-info.sql
@@ -1,0 +1,2 @@
+INSERT INTO soil_and_pimp_live.events (event_key, name, broadcast, schedule_change)
+     VALUES ('a', 'a', true, false), ('b', 'b', false, false);
